Correct live BVN URL
Live TV is not working at the moment. When you press play the UI will flicker for a second and then return to the list of channels.
I believe (not tested) this will fix the issue, or at least one of them, since the value returned by `live_jsonparser["hls"]` will already include the `http://` part.