Include time in video titles
Despite it being replay, BVN compiles a list of broadcasts for each day in a specific order. Not including something like the time in the title makes it impossible to distinguish certain broadcasts because you can have multiple items with the same title (news, etc). It would result in those items being listed together as sorting is being done based on title by default, making it look like a duplicate entry bug.

Sorting in descending order would be even better, but item.info['sorttitle'] with including an index before the title does not seem to be respected.