[PROJECT] Direct urls to github xmltv files.
The previous urls redirected with a 302 HTTP status. As it happend, urlquick caches all redirects with its default cache time, even when caching was explicitly disabled in the original request.

Because of this, we still end up with the old xml file when the hash has changed.

This commit implements the redirected url's.